@@ -1,6 +1,6 @@
 module geometrics
 
-go 1.18
+go 1.19
 
 require (
 	github.com/friendsofgo/errors v0.9.2
